kinfu_module merged into rgbd module
