--- modules/kinect_fusion/CMakeLists.txt
@@ -1,2 +1,0 @@
-set(the_description "Kinect Fusion implementation")
-ocv_define_module(kinect_fusion opencv_core opencv_imgproc OPTIONAL opencv_viz WRAP python)
