--- modules/kinect_fusion/README.md
@@ -1,15 +1,0 @@
-Kinect Fusion
-=============
-
-Implementation of KinectFusion algorithm
-Based on (kinfu-remake)[https://github.com/Nerei/kinfu_remake]
-
-Note that the KinectFusion algorithm was patented and its use may be restricted by following (but not limited to) list of patents:
-
-* _US20120196679A1_  Real-Time Camera Tracking Using Depth Maps
-* _US20120194644A1_  Mobile Camera Localization Using Depth Maps
-* _US20120194516A1_  Three-Dimensional Environment Reconstruction
-* _US8401225B2_  Moving object segmentation using depth images
-
-Since OpenCV's license imposes different restrictions on usage please consult a legal before using this algorithm any way.
-
