--- modules/kinect_fusion/include/opencv2/kinect_fusion.hpp
@@ -1,14 +1,0 @@
-// This file is part of OpenCV project.
-// It is subject to the license terms in the LICENSE file found in the top-level directory
-// of this distribution and at http://opencv.org/license.html
-
-// This code is also subject to the license terms in the LICENSE file found in this module's directory
-
-#include "opencv2/kinect_fusion/kinfu.hpp"
-
-/** @defgroup kinect_fusion Kinect Fusion
-
-KinectFusion 3d reconstruction algorithm implementation
-Based on (kinfu-remake)[https://github.com/Nerei/kinfu_remake]
-
-*/
