--- modules/kinect_fusion/src/precomp.hpp
@@ -1,119 +1,0 @@
-// This file is part of OpenCV project.
-// It is subject to the license terms in the LICENSE file found in the top-level directory
-// of this distribution and at http://opencv.org/license.html
-
-// This code is also subject to the license terms in the LICENSE file found in this module's directory
-
-#ifndef __OPENCV_PRECOMP_H__
-#define __OPENCV_PRECOMP_H__
-
-#include "opencv2/imgproc.hpp"
-#include "opencv2/kinect_fusion/kinfu.hpp"
-#include "opencv2/core/hal/intrin.hpp"
-
-// Print execution times of each block marked with ScopeTime
-#define PRINT_TIME 0
-
-#if PRINT_TIME
-#include <iostream>
-#endif
-
-namespace cv {
-namespace kinfu {
-
-typedef float depthType;
-
-const float qnan = std::numeric_limits<float>::quiet_NaN();
-const cv::Vec3f nan3(qnan, qnan, qnan);
-#if CV_SIMD128
-const cv::v_float32x4 nanv(qnan, qnan, qnan, qnan);
-#endif
-
-inline bool isNaN(cv::Point3f p)
-{
-    return (cvIsNaN(p.x) || cvIsNaN(p.y) || cvIsNaN(p.z));
-}
-
-#if CV_SIMD128
-static inline bool isNaN(const cv::v_float32x4& p)
-{
-    return cv::v_check_any(p != p);
-}
-#endif
-
-struct ScopeTime
-{
-    ScopeTime(std::string name_, bool _enablePrint = true);
-    ~ScopeTime();
-
-#if PRINT_TIME
-    static int nested;
-    const std::string name;
-    const bool enablePrint;
-    double start;
-#endif
-};
-
-/** @brief Camera intrinsics */
-struct Intr
-{
-    /** Reprojects screen point to camera space given z coord. */
-    struct Reprojector
-    {
-        Reprojector() {}
-        inline Reprojector(Intr intr)
-        {
-            fxinv = 1.f/intr.fx, fyinv = 1.f/intr.fy;
-            cx = intr.cx, cy = intr.cy;
-        }
-        template<typename T>
-        inline cv::Point3_<T> operator()(cv::Point3_<T> p) const
-        {
-            T x = p.z * (p.x - cx) * fxinv;
-            T y = p.z * (p.y - cy) * fyinv;
-            return cv::Point3_<T>(x, y, p.z);
-        }
-
-        float fxinv, fyinv, cx, cy;
-    };
-    /** Projects camera space vector onto screen */
-    struct Projector
-    {
-        inline Projector(Intr intr) : fx(intr.fx), fy(intr.fy), cx(intr.cx), cy(intr.cy) { }
-        template<typename T>
-        inline cv::Point_<T> operator()(cv::Point3_<T> p) const
-        {
-            T invz = T(1)/p.z;
-            T x = fx*(p.x*invz) + cx;
-            T y = fy*(p.y*invz) + cy;
-            return cv::Point_<T>(x, y);
-        }
-        template<typename T>
-        inline cv::Point_<T> operator()(cv::Point3_<T> p, cv::Point3_<T>& pixVec) const
-        {
-            T invz = T(1)/p.z;
-            pixVec = cv::Point3_<T>(p.x*invz, p.y*invz, 1);
-            T x = fx*pixVec.x + cx;
-            T y = fy*pixVec.y + cy;
-            return cv::Point_<T>(x, y);
-        }
-        float fx, fy, cx, cy;
-    };
-    Intr() : fx(), fy(), cx(), cy() { }
-    Intr(float _fx, float _fy, float _cx, float _cy) : fx(_fx), fy(_fy), cx(_cx), cy(_cy) { }
-    Intr(cv::Matx33f m) : fx(m(0, 0)), fy(m(1, 1)), cx(m(0, 2)), cy(m(1, 2)) { }
-    // scale intrinsics to pyramid level
-    inline Intr scale(int pyr) const
-    {
-        float factor = (1.f /(1 << pyr));
-        return Intr(fx*factor, fy*factor, cx*factor, cy*factor);
-    }
-    inline Reprojector makeReprojector() const { return Reprojector(*this); }
-    inline Projector   makeProjector()   const { return Projector(*this);   }
-
-    float fx, fy, cx, cy;
-};
-
-} // namespace kinfu
-} // namespace cv
-#endif
