--- modules/kinect_fusion/test/test_main.cpp
@@ -1,7 +1,0 @@
-// This file is part of OpenCV project.
-// It is subject to the license terms in the LICENSE file found in the top-level directory
-// of this distribution and at http://opencv.org/license.html.
-
-#include "test_precomp.hpp"
-
-CV_TEST_MAIN("cv")
