+++ modules/rgbd/CMakeLists.txt
@@ -1,2 +1,2 @@
 set(the_description "RGBD algorithms")
-ocv_define_module(rgbd opencv_core opencv_calib3d opencv_imgproc WRAP python)
+ocv_define_module(rgbd opencv_core opencv_calib3d opencv_imgproc OPTIONAL opencv_viz WRAP python)
