+++ modules/rgbd/samples/kinfu_demo.cpp
@@ -8,7 +8,7 @@
 #include <fstream>
 #include <opencv2/imgproc.hpp>
 #include <opencv2/highgui.hpp>
-#include <opencv2/kinect_fusion.hpp>
+#include <opencv2/rgbd/kinfu.hpp>
 
 #ifdef HAVE_OPENCV_VIZ
