+++ modules/rgbd/src/frame.hpp
@@ -8,6 +8,7 @@
 #define __OPENCV_KINFU_FRAME_H__
 
 #include "precomp.hpp"
+#include "utils.hpp"
 
 namespace cv {
