+++ modules/rgbd/src/odometry.cpp
@@ -34,6 +34,7 @@
  */
 
 #include "precomp.hpp"
+#include "fast_icp.hpp"
 
 #if defined(HAVE_EIGEN) && EIGEN_WORLD_VERSION == 3
 #define HAVE_EIGEN3_HERE
@@ -1095,6 +1096,8 @@ Ptr<Odometry> Odometry::create(const String & odometryType)
         return makePtr<ICPOdometry>();
     else if (odometryType == "RgbdICPOdometry")
         return makePtr<RgbdICPOdometry>();
+    else if (odometryType == "FastICPOdometry")
+        return makePtr<FastICPOdometry>();
     return Ptr<Odometry>();
 }
 
@@ -1441,6 +1444,131 @@ bool RgbdICPOdometry::computeImpl(const Ptr<OdometryFrame>& srcFrame, const Ptr<
 
 //
 
+using namespace cv::kinfu;
+
+FastICPOdometry::FastICPOdometry() :
+    maxDistDiff(DEFAULT_MAX_DEPTH_DIFF()),
+    angleThreshold((float)(30. * CV_PI / 180.)),
+    sigmaDepth(0.04f),
+    sigmaSpatial(4.5f),
+    kernelSize(7)
+{
+    setDefaultIterCounts(iterCounts);
+}
+
+FastICPOdometry::FastICPOdometry(const Mat& _cameraMatrix,
+                                 float _maxDistDiff,
+                                 float _angleThreshold,
+                                 float _sigmaDepth,
+                                 float _sigmaSpatial,
+                                 float _kernelSize,
+                                 const std::vector<int>& _iterCounts) :
+    maxDistDiff(_maxDistDiff),
+    angleThreshold(_angleThreshold),
+    sigmaDepth(_sigmaDepth),
+    sigmaSpatial(_sigmaSpatial),
+    kernelSize(_kernelSize),
+    iterCounts(Mat(_iterCounts).clone()),
+    cameraMatrix(_cameraMatrix)
+{
+    if(iterCounts.empty())
+        setDefaultIterCounts(iterCounts);
+}
+
+Ptr<FastICPOdometry> FastICPOdometry::create(const Mat& _cameraMatrix,
+                                             float _maxDistDiff,
+                                             float _angleThreshold,
+                                             float _sigmaDepth,
+                                             float _sigmaSpatial,
+                                             float _kernelSize,
+                                             const std::vector<int>& _iterCounts)
+{
+    return makePtr<FastICPOdometry>(_cameraMatrix, _maxDistDiff, _angleThreshold,
+                                   _sigmaDepth, _sigmaSpatial, _kernelSize, _iterCounts);
+}
+
+Size FastICPOdometry::prepareFrameCache(Ptr<OdometryFrame>& frame, int cacheType) const
+{
+    Odometry::prepareFrameCache(frame, cacheType);
+
+    if(frame->depth.empty())
+    {
+        if(!frame->pyramidDepth.empty())
+            frame->depth = frame->pyramidDepth[0];
+        else if(!frame->pyramidCloud.empty())
+        {
+            Mat cloud = frame->pyramidCloud[0];
+            std::vector<Mat> xyz;
+            split(cloud, xyz);
+            frame->depth = xyz[2];
+        }
+        else
+            CV_Error(Error::StsBadSize, "Depth or pyramidDepth or pyramidCloud have to be set.");
+    }
+    checkDepth(frame->depth, frame->depth.size());
+
+    // mask isn't used by FastICP
+
+    Ptr<FrameGenerator> fg = makeFrameGenerator(KinFu::Params::PlatformType::PLATFORM_CPU);
+    Ptr<FrameCPU> f = (*fg)().dynamicCast<FrameCPU>();
+    Intr intr(cameraMatrix);
+    iterCounts.total();
+    float depthFactor = 1.f; // user should rescale depth manually
+    (*fg)(f, frame->depth, intr, iterCounts.total(), depthFactor,
+          sigmaDepth, sigmaSpatial, kernelSize);
+
+    frame->pyramidCloud.clear();
+    frame->pyramidNormals.clear();
+    for(size_t i = 0; i < f->points.size(); i++)
+    {
+        frame->pyramidCloud.push_back(f->points[i]);
+        frame->pyramidNormals.push_back(f->normals[i]);
+    }
+
+    return frame->depth.size();
+}
+
+void FastICPOdometry::checkParams() const
+{
+    CV_Assert(cameraMatrix.size() == Size(3,3) &&
+              (cameraMatrix.type() == CV_32FC1 ||
+               cameraMatrix.type() == CV_64FC1));
+
+    CV_Assert(maxDistDiff > 0);
+    CV_Assert(angleThreshold > 0);
+    CV_Assert(sigmaDepth > 0 && sigmaSpatial > 0 && kernelSize > 0);
+}
+
+bool FastICPOdometry::computeImpl(const Ptr<OdometryFrame>& srcFrame,
+                                  const Ptr<OdometryFrame>& dstFrame,
+                                  OutputArray Rt, const Mat& /*initRt*/) const
+{
+    kinfu::Intr intr(cameraMatrix);
+    std::vector<int> iterations = iterCounts;
+    Ptr<kinfu::ICP> icp = kinfu::makeICP(kinfu::KinFu::Params::PlatformType::PLATFORM_CPU,
+                                         intr,
+                                         iterations,
+                                         angleThreshold,
+                                         maxDistDiff);
+
+    Affine3f transform;
+    Ptr<FrameCPU> srcF = makePtr<FrameCPU>(), dstF = makePtr<FrameCPU>();
+    for(size_t i = 0; i < srcFrame->pyramidCloud.size(); i++)
+    {
+        srcF-> points.push_back(srcFrame->pyramidCloud  [i]);
+        srcF->normals.push_back(srcFrame->pyramidNormals[i]);
+        dstF-> points.push_back(dstFrame->pyramidCloud  [i]);
+        dstF->normals.push_back(dstFrame->pyramidNormals[i]);
+    }
+    bool result = icp->estimateTransform(transform, srcF, dstF);
+
+    Rt.create(Size(4, 4), CV_64FC1);
+    Mat(Matx44d(transform.matrix)).copyTo(Rt.getMat());
+    return result;
+}
+
+//
+
 void
 warpFrame(const Mat& image, const Mat& depth, const Mat& mask,
           const Mat& Rt, const Mat& cameraMatrix, const Mat& distCoeff,
@@ -1453,5 +1581,5 @@ warpFrame(const Mat& image, const Mat& depth, const Mat& mask,
     else
         CV_Error(Error::StsBadArg, "Image has to be type of CV_8UC1 or CV_8UC3");
 }
-}
+} // namespace rgbd
 } // namespace cv
